% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{calc_accuracy}
\alias{calc_accuracy}
\title{Calculate Accuracy}
\usage{
calc_accuracy(tabble)
}
\arguments{
\item{tabble}{A frequency table created with \code{\link{table}}}
}
\value{
A tibble with the corresponding statistics
}
\description{
Calculates accuracy and related metrics.
}
\details{
Calculates accuracy, lower and upper bounds, the guessing rate and
p-value of the accuracy vs. the guessing rate. This function is called by
\code{confusion_matrix}, but if this is all you want, you can simply supply
the table to this function.
}
\examples{
\dontrun{
p = sample(letters[1:4], 250, replace = TRUE, prob = 1:4)
o = sample(letters[1:4], 250, replace = TRUE, prob = 1:4)
calc_accuracy(table(p, o))
}

}
\seealso{
\code{\link{binom.test}}
}
\author{
Michael Clark (see \href{https://github.com/m-clark/confusionMatrix}{m-clark/confusion_matrix}).
}
