% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam.R
\name{fam2ped}
\alias{fam2ped}
\title{Fam to pedigree}
\usage{
fam2ped(famfile, plotped = TRUE)
}
\arguments{
\item{famfile}{PLINK-formatted fam file}

\item{plotped}{Logical, TRUE if you want to create a PDF with pedigree drawings at the same location as the fam file with extension .pedigree.pdf}
}
\value{
A tibble with new listcol \code{ped} containing pedigrees from \code{kinship2::pedigree}.
}
\description{
Fam to pedigree
}
