% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd.R
\name{ibd2kin}
\alias{ibd2kin}
\title{Compute kinship coefficient from IBD segments}
\usage{
ibd2kin(.ibd_data, .map)
}
\arguments{
\item{.ibd_data}{Tibble with IBD segments created using the \link[skater]{read_ibd} function}

\item{.map}{Tibble with the genetic map data created using the \link[skater]{read_map} function}
}
\value{
Tibble with three columns:
\enumerate{
\item id1 (sample identifier 1)
\item id2 (sample identifier 2)
\item kinship (kinship coefficent derived from shared segments)
}
}
\description{
This function is used to retrieve a relatedness measure from IBD segments. The relatedness value returned is the kinship coefficent.

The input data should be pairwise IBD segments prepared via \link[skater]{read_ibd}. The function will internally loop over each chromosome, and use a specified genetic map to convert shared segments to genetic units. After doing so, the function converts the shared length to a kinship coefficient.
}
\references{
http://faculty.washington.edu/sguy/ibd_relatedness.html
}
