% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_degree}
\alias{infer_degree}
\title{Infer degree for a kinship coefficient}
\usage{
infer_degree(k, dibble)
}
\arguments{
\item{k}{Kinship coefficient (numeric, typically between 0 and .5, although KING can produce values <0).}

\item{dibble}{Degree tibble from \link[=dibble]{dibble}.}
}
\value{
Inferred degree, up to the maximum degree in \code{dibble} (anything more distant is \code{NA}, i.e., unrelated).
}
\description{
Infer degree for a kinship coefficient
}
\examples{
d3 <- dibble(3)
infer_degree(0.5, d3)
infer_degree(0.25, d3)
infer_degree(0.125, d3)
infer_degree(0.0625, d3)
infer_degree(0, d3)
infer_degree(-0.05, d3)
# Run on multiple kinship coefficients
k <- seq(.02, .5, .03)
# Not vectorized, this will not work:
# infer_degree(k, d3)
# This does:
purrr::map_int(k, infer_degree, d3)
tibble::tibble(k=k) \%>\% dplyr::mutate(degree=purrr::map_int(k, infer_degree, d3))
}
