% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate over segments}
\usage{
interpolate(ibd_bp, chromgpos)
}
\arguments{
\item{ibd_bp}{Base pair for the IBD segment over which to interpolate}

\item{chromgpos}{Genetic map data for a specific chromosome}
}
\value{
Numeric vector with the genetic distance shared at the segment.
}
\description{
This is an unexported helper used in in \link[skater]{ibd2kin}. The function interpolates over segments to apply genetic length to the segment. It is inspired by Python code distributed by the Browning lab (\href{http://faculty.washington.edu/sguy/ibd_relatedness.html}{documentation}).
}
\references{
http://faculty.washington.edu/sguy/ibd_relatedness.html
}
