% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam.R
\name{plot_pedigree}
\alias{plot_pedigree}
\title{Plot pedigree}
\usage{
plot_pedigree(ped, file = NULL, width = 10, height = 8)
}
\arguments{
\item{ped}{List of pedigree objects from \link{fam2ped}}

\item{file}{Output file path (must end in ".pdf")}

\item{width}{Width of output PDF}

\item{height}{Height of output PDF}
}
\description{
Plot pedigree
}
