% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_map}
\alias{read_map}
\title{Read genetic map file}
\usage{
read_map(file, source = "hapmap")
}
\arguments{
\item{file}{Input file path}

\item{source}{Source of the input file; must be one of \code{"hapmap"} or \code{"behrer"}; default is \code{"hapmap"}}
}
\value{
A tibble containing 3 columns:
\enumerate{
\item chr (chromosome)
\item value (genetic length within the physical position boundary)
\item bp (physical position boundary)
}
}
\description{
This function reads in the content from a genetic map file to translate physical distance to genetic units (i.e. cM).

The genetic map could come from different sources. One source is "hapmap", which is a map based on HapMap and distributed by the Browning Lab (\href{http://bochet.gcc.biostat.washington.edu/beagle/genetic_maps/}{documentation}). If this map file is used, the non-sex chromosomes can be downloaded and concatentated to a single file as follows:\preformatted{wget http://bochet.gcc.biostat.washington.edu/beagle/genetic_maps/plink.GRCh37.map.zip
unzip plink.GRCh37.map.zip
cat *chr[0-9]*GRCh37.map | sort -k1,1 -k4,4 --numeric-sort > plink.allchr.GRCh37.map
}

Another source is a sex-specific map ("behrer") originally published by Behrer et al and recommended by the developers of \code{ped-sim} (\href{https://github.com/williamslab/ped-sim#map-file}{documentation}). To retrieve and prep this map file:\preformatted{wget https://github.com/cbherer/Bherer_etal_SexualDimorphismRecombination/raw/master/Refined_genetic_map_b37.tar.gz
tar xvzf Refined_genetic_map_b37.tar.gz
printf "#chr\\tpos\\tmale_cM\\tfemale_cM\\n" > refined_mf.simmap
for chr in \{1..22\}; do
paste Refined_genetic_map_b37/male_chr$chr.txt Refined_genetic_map_b37/female_chr$chr.txt \\
| awk -v OFS="\\t" 'NR > 1 && $2 == $6 \{print $1,$2,$4,$8\}' \\
| sed 's/^chr//' >> refined_mf.simmap;
done
}

Note that if the "behrer" map file is specified as source, then the sex-specific cM lengths will be averaged.
}
\references{
\url{http://zzz.bwh.harvard.edu/plink/data.shtml#map}

\url{http://bochet.gcc.biostat.washington.edu/beagle/genetic_maps/}

\url{https://github.com/williamslab/ped-sim#map-file}

\url{http://dx.doi.org/10.1038/ncomms14994}
}
