% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skater-package.R
\docType{package}
\name{skater-package}
\alias{skater}
\alias{skater-package}
\title{skater: Utilities for SNP-Based Kinship Analysis}
\description{
Utilities for SNP-based kinship analysis testing and evaluation (SKATE).
}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

Authors:
\itemize{
  \item Matthew Scholz (\href{https://orcid.org/0000-0003-3686-1227}{ORCID})
  \item VP Nagraj (\href{https://orcid.org/0000-0003-0060-566X}{ORCID})
}

}
\keyword{internal}
