% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam.R
\name{fam2ped}
\alias{fam2ped}
\title{Fam to pedigree}
\usage{
fam2ped(famfile, plotped = TRUE)
}
\arguments{
\item{famfile}{PLINK-formatted fam file}

\item{plotped}{Logical, TRUE if you want to create a PDF with pedigree drawings at the same location as the fam file with extension .pedigree.pdf}
}
\value{
A tibble with new listcol \code{ped} containing pedigrees from \code{kinship2::pedigree}.
}
\description{
Converts a \href{https://www.cog-genomics.org/plink/1.9/formats#fam}{PLINK-formatted fam file} to a pedigree object using \link[kinship2:pedigree]{kinship2::pedigree}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
