% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam.R
\name{ped2kinpair}
\alias{ped2kinpair}
\title{Pedigree to pairwise kinship}
\usage{
ped2kinpair(ped)
}
\arguments{
\item{ped}{A "pedigree" class object from \link[skater]{fam2ped}.}
}
\value{
A tibble containing all pairwise kinship coefficients from the input pedigree.
}
\description{
Pedigree to pairwise kinship
}
