% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_map}
\alias{read_map}
\title{Read genetic map file}
\usage{
read_map(file, source = "hapmap")
}
\arguments{
\item{file}{Input file path}

\item{source}{Source of the input file; currently only \code{"hapmap"} supported (and set to default)}
}
\value{
A tibble containing 3 columns:
\enumerate{
\item chr (chromosome)
\item value (genetic length within the physical position boundary)
\item bp (physical position boundary)
}
}
\description{
This function reads in the content from a genetic map file to translate physical distance to genetic units (i.e. cM).

The genetic map could come from different sources. However, the only is the map file based on HapMap and distributed by the Browning Lab (\href{http://bochet.gcc.biostat.washington.edu/beagle/genetic_maps/}{documentation}). If this map file is used, the non-sex chromosomes should be concatentated to a single file as follows:

\verb{cat *chr[0-9]*GRCh37.map | sort -k1,1 -k4,4 --numeric-sort > plink.allchr.GRCh37.map}
}
\references{
\url{http://zzz.bwh.harvard.edu/plink/data.shtml#map}

\url{http://bochet.gcc.biostat.washington.edu/beagle/genetic_maps/}
}
