% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skater-package.R
\docType{package}
\name{skater-package}
\alias{skater}
\alias{skater-package}
\title{skater: Utilities for SNP-based kinship analysis}
\description{
Utilities for SNP-based kinship analysis.
}
\author{
\strong{Maintainer}: Stephen Turner \email{sturner@signaturescience.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

}
\keyword{internal}
