% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skater-package.R
\docType{package}
\name{skater-package}
\alias{skater}
\alias{skater-package}
\title{skater: Utilities for SNP-Based Kinship Analysis}
\description{
Utilities for single nucleotide polymorphism (SNP) based kinship analysis testing and evaluation. The 'skater' package contains functions for importing, parsing, and analyzing pedigree data, performing relationship degree inference, benchmarking relationship degree classification, and summarizing identity by descent (IBD) segment data. Package functions and methods are described in Turner et al. (2021) "skater: An R package for SNP-based Kinship Analysis, Testing, and Evaluation" \doi{10.1101/2021.07.21.453083}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/signaturescience/skater}
  \item Report bugs at \url{https://github.com/signaturescience/skater/issues}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

Authors:
\itemize{
  \item Matthew Scholz (\href{https://orcid.org/0000-0003-3686-1227}{ORCID})
  \item VP Nagraj (\href{https://orcid.org/0000-0003-0060-566X}{ORCID})
}

Other contributors:
\itemize{
  \item Signature Science, LLC. [copyright holder]
}

}
\keyword{internal}
